+++ src/App.jsx
@@ -1,18 +1,13 @@
-const todoList = [
-    { id: 1, title: "Complete lesson1" },
-    { id: 2, title: "Complete lesson1" },
-    { id: 3, title: "Complete lesson3" },
-];
+import "./App.css";
+import AddTodoForm from "./AddTodoForm";
+import TodoList from "./TodoList";
 
 function App() {
     return (
         <div className="App">
             <h1>My Todo List</h1>
-            <ul>
-                {todoList.map((item) => (
-                    <li key={item.id}>{item.title}</li>
-                ))}
-            </ul>
+            <TodoList />
+            <AddTodoForm />
         </div>
     );
 }
